% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_loglik.R
\name{.bf_loglik_coloc}
\alias{.bf_loglik_coloc}
\title{Multi-INTACT EM algorithm log-likelihood function.}
\usage{
.bf_loglik_coloc(w, bf, fp_coloc)
}
\arguments{
\item{w}{A vector of prior parameter estimates.}

\item{bf}{A vector of Bayes factors.}

\item{fp_coloc}{A vector of aggregated colocalization evidence. The default
is the max of all GLCPs, truncated at 0.05.}
}
\value{
A scalar log likelihood.
}
\description{
Multi-INTACT EM algorithm log-likelihood function.
}
