% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_weighted_sum.R
\name{.bf_weighted_sum_coloc}
\alias{.bf_weighted_sum_coloc}
\title{Helper function for EM algorithm to compute weighted sum of Bayes factors.}
\usage{
.bf_weighted_sum_coloc(w, bf, fp_coloc, i)
}
\arguments{
\item{w}{A vector of prior parameter estimates.}

\item{bf}{A vector of Bayes factors.}

\item{fp_coloc}{A vector of aggregated colocalization evidence. The default
is the max of all GLCPs, truncated at 0.05.}

\item{i}{The index for the ith gene.}
}
\value{
A scalar weighted sum of Bayes factors
}
\description{
Helper function for EM algorithm to compute weighted sum of Bayes factors.
}
