% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_est.R
\name{.em_est}
\alias{.em_est}
\title{Compute gene set enrichment estimates.}
\usage{
.em_est(pprobs, d_vec)
}
\arguments{
\item{pprobs}{A vector of posterior probabilities for each gene estimated
from the
intact function. Gene order should match d_vec.}

\item{d_vec}{A vector of gene set annotations for the genes of interest.
Entries should be integer(1) if the gene is annotated and integer(0)
otherwise.}
}
\value{
Maximum likelihood estimates for alpha0 and alpha1; convergence
indicator.
}
\description{
Compute gene set enrichment estimates.
}
\examples{
data(simdat)
.em_est(d_vec = sample(c(0,1),1197,replace=TRUE),
pprobs = intact(GLCP_vec=simdat$GLCP,
prior_fun=linear, z_vec = simdat$TWAS_z, t = 0.05))
}
