% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{.logistic_loglik}
\alias{.logistic_loglik}
\title{A log likelihood function for the expectation-maximization algorithm.
Used as an argument for objfn in the squarem function.}
\usage{
.logistic_loglik(alpha, d_vec, pprobs)
}
\arguments{
\item{alpha}{A vector containing the current estimates of the enrichment
parameters
alpha0 and alpha1.}

\item{d_vec}{A vector of gene set annotations for the genes of interest.
Entries should be integer(1) if the gene is annotated and integer(0)
otherwise.}

\item{pprobs}{A vector of posterior probabilities for each gene estimated
from the
intact function. Gene order should match d_vec.}
}
\value{
Log likelihood evaluated at the current estimates of alpha0 and
alpha1.
}
\description{
A log likelihood function for the expectation-maximization algorithm.
Used as an argument for objfn in the squarem function.
}
