% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_em_posterior.R
\name{.multi_em_posteriors}
\alias{.multi_em_posteriors}
\title{Compute gene product relevance probabilities using prior parameter estimates
and Bayes factors.}
\usage{
.multi_em_posteriors(w, bf, fp_coloc)
}
\arguments{
\item{w}{A vector of prior parameter estimates.}

\item{bf}{A vector of Bayes factors.}

\item{fp_coloc}{A vector of aggregated colocalization evidence. The default
is the max of all GLCPs, truncated at 0.05.}
}
\value{
A vector of posterior probabilities.
}
\description{
Compute gene product relevance probabilities using prior parameter estimates
and Bayes factors.
}
