% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_bf.R
\name{.wakefield_bf_z_ln}
\alias{.wakefield_bf_z_ln}
\title{A function to compute log Bayes factors from z-statistics using the
Wakefield formula}
\usage{
.wakefield_bf_z_ln(z_vec, K = c(1, 2, 4, 8, 16))
}
\arguments{
\item{z_vec}{A vector of TWAS scan z-scores for the genes of interest.}

\item{K}{A vector of values over which Bayesian model averaging is performed.}
}
\value{
log Bayes factors
}
\description{
A function to compute log Bayes factors from z-statistics using the
Wakefield formula
}
