% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_se.R
\name{enrich_bootstrap_se}
\alias{enrich_bootstrap_se}
\title{Compute bootstrap standard errors for alpha MLEs.}
\usage{
enrich_bootstrap_se(pprobs, d_vec, reps = 100)
}
\arguments{
\item{pprobs}{A vector of posterior probabilities for each gene estimated
from
the evid_int function. Gene order should match d_vec.}

\item{d_vec}{A vector of gene set annotations for the genes of interest.
Entries should be integer(1) if the gene is annotated and integer(0)
otherwise.}

\item{reps}{Number of bootstrap samples.}
}
\value{
Boostrap standard errors of MLEs for alpha0 and alpha1.
}
\description{
Compute bootstrap standard errors for alpha MLEs.
}
\examples{
enrich_bootstrap_se(d_vec = sample(c(0,1),1197,replace=TRUE),
pprobs = intact(GLCP_vec=simdat$GLCP,prior_fun=linear, z_vec = simdat$TWAS_z,
 t = 0.05))
}
