% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exprwt_sumstats.R
\docType{data}
\name{exprwt_sumstats}
\alias{exprwt_sumstats}
\title{TWAS weights for a simulated gene.}
\format{
A vector of length 1500.
}
\usage{
exprwt_sumstats
}
\value{
A vector of length 1500.
}
\description{
TWAS weights 1500 SNPs in a simulated data set. This data is
included to demonstrate how to compute a multivariate Wald statistic from
summary-level data.
}
\examples{
data(exprwt_sumstats)
}
\keyword{datasets}
