% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{fdr_rst}
\alias{fdr_rst}
\title{Bayesian FDR control for INTACT output}
\usage{
fdr_rst(posterior, alpha = 0.05)
}
\arguments{
\item{posterior}{A vector of posterior probabilities for each gene estimated
from the
intact function.}

\item{alpha}{A numeric target FDR control level.}
}
\value{
An n x 2 data frame where the first column is the inputted
posterior probabilities, and the second is a Boolean vector denoting
significance at the specified target control level.
}
\description{
Bayesian FDR control for INTACT output
}
\examples{
data(simdat)
fdr_rst(simdat$GLCP)
}
