\name{intact}
\alias{intact}
\alias{intact}
\title{INTACT: Integration of longitudinal physical activity data from multiple sources}

\description{
The \pkg{intact} package implements the INTACT method for harmonizing longitudinal 
physical activity features collected from multiple sources. It provides 
tools to adjust for batch effects while preserving biologically and behaviorally 
relevant variation, enabling integrated downstream analysis.
}

\author{
Jingru Zhang \cr
Maintainer: Jingru Zhang \email{jr_zhang@fudan.edu.cn}
}

\references{
Zhang, J., Cui, E., Li, H., & Shou, H. (2025). 
\emph{INTACT: A method for integration of longitudinal physical activity data from multiple sources}. 
bioRxiv.
}

\seealso{
\code{\link{INTACT}}, \code{\link{INTACTforXW}}, \code{\link{fastmfpca}}
}
