% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intact.R
\name{intact}
\alias{intact}
\title{Compute the posterior probability that a gene may be causal, given a gene's TWAS scan z-score and colocalization probability.}
\usage{
intact(GLCP_vec, prior_fun, z_vec, t = NULL, K = c(1, 2, 4, 8, 16))
}
\arguments{
\item{GLCP_vec}{A vector of colocalization probabilities for the genes of interest}

\item{prior_fun}{A function to transform a colocalization probability into a prior.
Options are linear, step, expit, and hybrid.}

\item{z_vec}{A vector of TWAS scan z-scores for the genes of interest. The order of
genes must match GLCP_vec.}

\item{t}{A hard threshold for the GLCP. Values below this number will be shrunk to zero.
This argument is used in the user-specified prior function.}

\item{K}{A vector of values over which Bayesian model averaging is performed.}
}
\value{
The vector of posteriors.
}
\description{
Compute the posterior probability that a gene may be causal, given a gene's TWAS scan z-score and colocalization probability.
}
\examples{
intact(GLCP_vec=simdat$GLCP, prior_fun=linear, z_vec = simdat$TWAS_z,t = 0.05)
}
