% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intactGSE.R
\name{intactGSE}
\alias{intactGSE}
\title{Perform gene set enrichment estimation and inference, given TWAS scan z-scores and colocalization probabilities.}
\usage{
intactGSE(
  gene_data,
  prior_fun,
  gene_sets,
  t = 0.05,
  sig_lev = 0.05,
  SE_type = "NDS",
  boot_rep = NULL
)
}
\arguments{
\item{gene_data}{A dataframe containing gene names and corresponding colocalization
probabilities and TWAS z-scores for each gene. Column names should be "gene", "GLCP",
and "TWAS_z'}

\item{prior_fun}{A function to transform a colocalization probability into a prior.
Options are linear, step, expit, and hybrid.}

\item{gene_sets}{A named list of gene sets for which enrichment is to be estimated.
List items should be character vectors of gene IDs. Gene ID format should match the
 gene column in gene_data.}

\item{t}{A hard threshold for the GLCP. Values below this number will be shrunk to zero.
This argument is used in the user-specified prior function.}

\item{sig_lev}{A significance threshold for gene set enrichment hypothesis testing.}

\item{SE_type}{A method to compute standard errors of the gene set enrichment estimates.
Possible methods are "profile_likelihood" and "bootstrap."}

\item{boot_rep}{Number of bootstrap samples.}
}
\value{
A data frame with the alpha1 estimate, standard error, z-score, p-value,
(1-sig_lev)\% CI limits, and convergence indicator for each gene set in gene_sets.
}
\description{
Perform gene set enrichment estimation and inference, given TWAS scan z-scores and colocalization probabilities.
}
\examples{
intactGSE(gene_data = simdat,prior_fun = linear,gene_sets = gene_set_list)
}
