% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_sumstats.R
\docType{data}
\name{ld_sumstats}
\alias{ld_sumstats}
\title{LD correlation matrix from a simulated data set.}
\format{
A matrix with 1500 rows and 1500 columns.
}
\usage{
ld_sumstats
}
\value{
A matrix with 1500 rows and 1500 columns.
}
\description{
A data set containing correlation data for 1500 cis-SNPs for a simulated
target gene. This data is included to demonstrate how to compute a
multivariate Wald statistic from summary-level data.
}
\examples{
data(ld_sumstats)
}
\keyword{datasets}
