% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixptfn2.R
\name{logistic_em_nopseudo}
\alias{logistic_em_nopseudo}
\title{Similar to logistic_em(), but does not use pseudocounts to stablize the algorithm.}
\usage{
logistic_em_nopseudo(d_vec, pprobs, alpha)
}
\arguments{
\item{d_vec}{A vector of gene set annotations for the genes of interest.
Entries should be integer(1) if the gene is annotated and integer(0) otherwise.}

\item{pprobs}{A vector of posterior probabilities for each gene estimated from the
intact function. Gene order should match d_vec.}

\item{alpha}{A vector containing the current estimates of the enrichment parameters
alpha0 and alpha1.}
}
\value{
Updated estimates of alpha0 and alpha1.
}
\description{
Similar to logistic_em(), but does not use pseudocounts to stablize the algorithm.
}
\examples{
logistic_em_nopseudo(d_vec = sample(c(0,1),1197,replace=TRUE),
pprobs = intact(GLCP_vec=simdat$GLCP, prior_fun=linear, z_vec = simdat$TWAS_z, t = 0.05),
alpha = c(0,0))
}
