% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{logistic_loglik}
\alias{logistic_loglik}
\title{A log likelihood function for the expectation-maximization algorithm.
Used as an argument for objfn in the squarem function.}
\usage{
logistic_loglik(alpha, d_vec, pprobs)
}
\arguments{
\item{alpha}{A vector containing the current estimates of the enrichment
parameters
alpha0 and alpha1.}

\item{d_vec}{A vector of gene set annotations for the genes of interest.
Entries should be integer(1) if the gene is annotated and integer(0)
otherwise.}

\item{pprobs}{A vector of posterior probabilities for each gene estimated
from the
intact function. Gene order should match d_vec.}
}
\value{
Log likelihood evaluated at the current estimates of alpha0 and
alpha1.
}
\description{
A log likelihood function for the expectation-maximization algorithm.
Used as an argument for objfn in the squarem function.
}
\examples{
data(simdat)
logistic_loglik(d_vec = sample(c(0,1),1197,replace=TRUE),
pprobs = intact(GLCP_vec=simdat$GLCP, prior_fun=linear,
z_vec = simdat$TWAS_z, t = 0.05),
alpha = c(0,0))
}
