% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_simdat.R
\docType{data}
\name{multi_simdat}
\alias{multi_simdat}
\title{Simulated TWAS, PWAS, and pairwise colocalization summary data.}
\format{
A data frame with 1197 rows and 6 variables:
\describe{
  \item{gene}{gene Ensembl ID}
  \item{GLCP_1}{Pairwise colocalization probability for the complex trait and
  gene expression levels.}
  \item{GLCP_2}{Pairwise colocalization probability for the complex trait and
  encoded protein levels.}
  \item{z_1}{TWAS z-score.}
  \item{z_2}{PWAS z-score.}
  \item{chisq}{Multivariate Wald chi-square statistic from the regression
  of the complex trait on the predicted expression and protein levels of the
  target gene.}
  ...
}
}
\usage{
multi_simdat
}
\value{
A data frame with 1197 rows and 6 variables:
}
\description{
A data set containing pairwise fastENLOC GLCPs, TWAS and PWAS z-scores, and
multivariate Wald chi-square statistics for 1197 simulated genes.
}
\examples{
data(multi_simdat)
}
\keyword{datasets}
