% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi1.R
\name{pi1_fun}
\alias{pi1_fun}
\title{Estimate pi1 from TWAS scan z-scores.}
\usage{
pi1_fun(z_vec, lambda = 0.5)
}
\arguments{
\item{z_vec}{A vector of TWAS scan z-scores.}

\item{lambda}{A value between 0 and 1. The density of TWAS scan z-scores
should be
flat at lambda. Set to 0.5 as default.}
}
\value{
A scalar estimate for pi1.
}
\description{
Estimate pi1 from TWAS scan z-scores.
}
\examples{
pi1_fun(simdat$TWAS_z)
}
