% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protwt_sumstats.R
\docType{data}
\name{protwt_sumstats}
\alias{protwt_sumstats}
\title{PWAS weights for a simulated gene.}
\format{
A vector of length 1500.
}
\usage{
protwt_sumstats
}
\value{
A vector of length 1500.
}
\description{
PWAS weights 1500 SNPs in a simulated data set. This data is
included to demonstrate how to compute a multivariate Wald statistic from
summary-level data.
}
\examples{
data(protwt_sumstats)
}
\keyword{datasets}
