% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{simdat}
\alias{simdat}
\title{Simulated TWAS and colocalization summary data.}
\format{
A data frame with 1197 rows and 3 variables:
\describe{
  \item{gene}{gene Ensembl ID}
  \item{GLCP}{colocalization probability}
  \item{TWAS_z}{TWAS z-score}
  ...
}
}
\usage{
simdat
}
\description{
A dataset containing GLCP and TWAS z-score for 1197 simulated genes.
}
\keyword{datasets}
