% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sumstats.R
\docType{data}
\name{z_sumstats}
\alias{z_sumstats}
\title{TWAS and PWAS z-score for a simulated gene.}
\format{
A 2-vector.
}
\usage{
z_sumstats
}
\value{
A 2-vector.
}
\description{
A 2-vector for which the first entry is the TWAS z-score, and the second
entry is the PWAS z-score for a simulated gene. This data is included to
demonstrate how to compute a multivariate Wald statistic from summary-level
data.
}
\examples{
data(z_sumstats)
}
\keyword{datasets}
