% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_set_list.R
\docType{data}
\name{gene_set_list}
\alias{gene_set_list}
\title{Simulated gene set list.}
\format{
A list with two items:
\describe{
  \item{gene set 1}{gene set with 503 gene members. Significantly enriched
  in simdat.}
  \item{gene set 2}{gene set with 200 members.}
  ...
}
}
\usage{
gene_set_list
}
\description{
A list object containing two elements. Each is a character list of gene
names.
}
\examples{
data(gene_set_list)
}
\keyword{datasets}
